+++ federation-integration-testsuite-js/src/utils/index.ts
@@ -1,0 +1,1 @@
+export { gql } from './gql';