+++ federation-js/src/composition/validate/postComposition/keyFieldsMissingOnBase.ts
@@ -1,5 +1,5 @@
 import { isObjectType, FieldNode, GraphQLError } from 'graphql';
-import { logServiceAndType, errorWithCode, getFederationMetadata } from '../../utils';
+import { logServiceAndType, errorWithCode, getFederationMetadata, findTypeNodeInServiceList, findSelectionSetOnNode, isDirectiveDefinitionNode, printFieldSet } from '../../utils';
 import { PostCompositionValidator } from '.';
 
 /**
@@ -7,6 +7,7 @@ import { PostCompositionValidator } from '.';
  */
 export const keyFieldsMissingOnBase: PostCompositionValidator = ({
   schema,
+  serviceList,
 }) => {
   const errors: GraphQLError[] = [];
 
@@ -30,6 +31,10 @@ export const keyFieldsMissingOnBase: PostCompositionValidator = ({
             // NOTE: We don't need to warn if there is no matching field.
             // keyFieldsSelectInvalidType already does that :)
             if (matchingField) {
+              const typeNode = findTypeNodeInServiceList(typeName, serviceName, serviceList);
+              const selectionSetNode = !isDirectiveDefinitionNode(typeNode) ?
+                findSelectionSetOnNode(typeNode, 'key', printFieldSet(selectionSet)) : undefined;
+
               const fieldFederationMetadata = getFederationMetadata(matchingField);
               // warn if not from base type OR IF IT WAS OVERWITTEN
               if (fieldFederationMetadata?.serviceName) {
@@ -38,6 +43,7 @@ export const keyFieldsMissingOnBase: PostCompositionValidator = ({
                     'KEY_FIELDS_MISSING_ON_BASE',
                     logServiceAndType(serviceName, typeName) +
                       `A @key selects ${name}, but ${typeName}.${name} was either created or overwritten by ${fieldFederationMetadata.serviceName}, not ${serviceName}`,
+                    selectionSetNode,
                   ),
                 );
               }