+++ federation-js/src/composition/validate/preComposition/__tests__/keyFieldsMissingExternal.test.ts
@@ -1,6 +1,8 @@
-import gql from 'graphql-tag';
 import { keyFieldsMissingExternal as validateKeyFieldsMissingExternal } from '../';
-import { graphqlErrorSerializer } from 'apollo-federation-integration-testsuite';
+import {
+  gql,
+  graphqlErrorSerializer,
+} from 'apollo-federation-integration-testsuite';
 
 expect.addSnapshotSerializer(graphqlErrorSerializer);
 
@@ -105,6 +107,12 @@ describe('keyFieldsMissingExternal', () => {
       Array [
         Object {
           "code": "KEY_FIELDS_MISSING_EXTERNAL",
+          "locations": Array [
+            Object {
+              "column": 3,
+              "line": 3,
+            },
+          ],
           "message": "[serviceA] Product -> A @key directive specifies the \`sku\` field which has no matching @external field.",
         },
       ]
@@ -133,6 +141,12 @@ describe('keyFieldsMissingExternal', () => {
       Array [
         Object {
           "code": "KEY_FIELDS_MISSING_EXTERNAL",
+          "locations": Array [
+            Object {
+              "column": 35,
+              "line": 1,
+            },
+          ],
           "message": "[serviceA] Product -> A @key directive specifies a field which is not found in this service. Add a field to this type with @external.",
         },
       ]
@@ -165,14 +179,32 @@ describe('keyFieldsMissingExternal', () => {
       Array [
         Object {
           "code": "KEY_FIELDS_MISSING_EXTERNAL",
+          "locations": Array [
+            Object {
+              "column": 3,
+              "line": 8,
+            },
+          ],
           "message": "[serviceA] Model -> A @key directive specifies the \`name\` field which has no matching @external field.",
         },
         Object {
           "code": "KEY_FIELDS_MISSING_EXTERNAL",
+          "locations": Array [
+            Object {
+              "column": 3,
+              "line": 9,
+            },
+          ],
           "message": "[serviceA] Model -> A @key directive specifies the \`kit\` field which has no matching @external field.",
         },
         Object {
           "code": "KEY_FIELDS_MISSING_EXTERNAL",
+          "locations": Array [
+            Object {
+              "column": 3,
+              "line": 13,
+            },
+          ],
           "message": "[serviceA] Kit -> A @key directive specifies the \`upc\` field which has no matching @external field.",
         },
       ]