+++ federation-js/src/composition/validate/sdl/__tests__/possibleTypeExtensions.test.ts
@@ -6,11 +6,11 @@ import {
   extendSchema,
 } from 'graphql';
 import { validateSDL } from 'graphql/validation/validate';
-import gql from 'graphql-tag';
 import { buildMapsFromServiceList } from '../../../compose';
 import {
   typeSerializer,
   graphqlErrorSerializer,
+  gql,
 } from 'apollo-federation-integration-testsuite';
 import federationDirectives from '../../../../directives';
 import { ServiceDefinition } from '../../../types';
@@ -106,6 +106,12 @@ describe('PossibleTypeExtensionsType', () => {
       Array [
         Object {
           "code": "EXTENSION_WITH_NO_BASE",
+          "locations": Array [
+            Object {
+              "column": 1,
+              "line": 2,
+            },
+          ],
           "message": "[serviceA] Product -> \`Product\` is an extension type, but \`Product\` is not defined in any service",
         },
       ]
@@ -140,13 +146,19 @@ describe('PossibleTypeExtensionsType', () => {
     schema = extendSchema(schema, definitions, { assumeValidSDL: true });
     errors.push(...validateSDL(extensions, schema, [PossibleTypeExtensions]));
     expect(errors).toMatchInlineSnapshot(`
-            Array [
-              Object {
-                "code": "EXTENSION_OF_WRONG_KIND",
-                "message": "[serviceA] Product -> \`Product\` was originally defined as a InputObjectTypeDefinition and can only be extended by a InputObjectTypeExtension. serviceA defines Product as a ObjectTypeExtension",
-              },
-            ]
-        `);
+      Array [
+        Object {
+          "code": "EXTENSION_OF_WRONG_KIND",
+          "locations": Array [
+            Object {
+              "column": 1,
+              "line": 2,
+            },
+          ],
+          "message": "[serviceA] Product -> \`Product\` was originally defined as a InputObjectTypeDefinition and can only be extended by a InputObjectTypeExtension. serviceA defines Product as a ObjectTypeExtension",
+        },
+      ]
+    `);
   });
 
   it('does not error', () => {